refactor: convert price from double to long for precision

+++ backend/src/main/java/com/swiftcart/swiftcart/features/cart/CartResponse.java
@@ -5,5 +5,5 @@ import java.util.List;
 public record CartResponse(
     Long id,
     List<CartItemResponse> items,
-    double totalPrice
+    long totalPrice
 ) {}
